Initial version of Transifex-2-Transifex
Particularly for language team not at 100%, translations may often be
done for the 'development' version and for each of the 'release'
versions of concrete5.
tx2tx.html, via a JavaScript library, fetches the Transifex
translations for a language and for two resources (eg 'core-dev' and
'core-562'), and compares them checking for missing or different
translations.
tx2tx also allows to copy these different translations.

At the time of writing, Transifex does not send the
Access-Control-Allow-Origin http header, so you have to start Chrome
with the parameter --disable-web-security (see http://goo.gl/SnFiDJ )